+++ src/sqlforcql/db.clj
@@ -12,11 +12,12 @@
 
 (defn connect-to-db
   "Connects to a Cassandra keyspace on a particular ip-address, with the given username and password."
-  [ip-address username password keyspace]
+  [ip-address port username password keyspace]
   (let [cluster (alia/cluster {:contact-points [ip-address]
-                               :credentials {:user username :password password}})
+                               :port           port
+                               :credentials    {:user username :password password}})
         session (alia/connect cluster)]
-    (debug "Connecting to" keyspace "on" ip-address)
+    (debug "Connecting to" keyspace "on" (str ip-address ":" port))
     (get-db-map cluster session keyspace)))
 
 (defn disconnect-from-db