+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [0.3.7] - 2021-05-23
+### Changed
+- In insert statement generation, escape single quotes for values of map<text,text>, set<text> columns.
+- Accept port number when connecting to Cassandra.
+
 ## [0.3.6] - 2021-05-17
 ### Changed
 - In insert statement generation, use the keyspace name along with the table name.